Fix compatibility with wp-graphql-gutenberg-acf
Simply casting the $attributes array to an object fails with ACF Pro's
field groups, as there's another nested array. Using JSON decode makes
sure that nested arrays are converted to objects as well

There is also an issue with the HTML loader failing if there's no
innerHTML defined in the block. HTML-less blocks are very useful for
WordPress as a headless CMS with WPGraphQL.